pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "ronny-langraph"
dynamic = ["version"]
description = 'experimenting with langraph'
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
keywords = []
authors = [
  { name = "Ronny Pfannschmidt", email = "opensource@ronnypfannschmidt.de" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
]
dependencies = [
  "langgraph-cli",
  "langchain-git",
  "langchain-addons" ,
  "langchain-ollama",
  "rich",
]

[project.urls]
Documentation = "https://github.com/Ronny Pfannschmidt/ronny-langraph#readme"
Issues = "https://github.com/Ronny Pfannschmidt/ronny-langraph/issues"
Source = "https://github.com/Ronny Pfannschmidt/ronny-langraph"


[dependency-groups]
test  = [
  "pytest>8.3",
  "pyright>=1.1.401",
  "mypy>=1.15.0",
]
dev = ["langgraph-cli[inmem]", ]
typing = []

[tool.uv]
default-groups="all"

[tool.hatch.version]
source = "vcs"


[tool.coverage.run]
source_pkgs = ["ronny_langraph", "tests"]
branch = true
parallel = true
omit = [
  "src/ronny_langraph/__about__.py",
]

[tool.coverage.paths]
ronny_langraph = ["src/ronny_langraph", "*/ronny-langraph/src/ronny_langraph"]
tests = ["tests", "*/ronny-langraph/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
